// reactjs/projects/ht-management/src/Members/PBT/pages/Components/DefaultLayout/HomeUser/type.ts
  
  export interface RoomType {
    roomName: string;
    roomDescription: string;
    roomImage: string;
    roomType:string;
  }
  export interface ServiceType {
    nameService: string;
    serviceDescription: string;
  }
  export interface StatusType {
    img1: string;
    img2: string;
    img3: string;
    nameStatus: string;
    introduce: string;
  }